Pivot origin selection by channel
Channels now dictate what origin is available. That way we can upgrade a
charm via deploy bundle that has a different channel. This not only
works for charmhub charms, but also charmstore charms. Previous
implementations of bundle deploy never took into account channels and
therefore I'm unsure how you would ever upgrade based on that. I presume
that because the revision of the charm url had changed, then that would
trigger an upgrade?

As a requirement for charmstore charm tests we should ensure that we
didn't break any backwards compatible code. For charmhub charms the code
ensure that we can detect a new revision and then use that revision when
upgrading.