Allow channels to be used for upgrades
When re-deploying a bundle again, against the same applications,
essentially performing an upgrade we should also take into account the
channel.

Previously channel was only viewed when deploying first time around. The
following also takes this into account when re-deploying. It will warn
you to use force in certain situations to ensure that you're doing
something that isn't exactly supported (swapping channels etc).

Most of the code here isn't actually complicated, it wired in the new
library updates from bundlechanges. The code essentially does another
request to the store when it notices an update to ensure that we know
what revision we can move to. All this code is super accessible and
because of the reusability of resolve charm it makes it nice and easy to
call when performing a bundle upgrade.

There could be more tests as always, but we've covered most of the
cases.