Better help message for existing applications
When deploying an existing application again via `juju deploy` the help
message is a bit sad, as it doesn't explain how you get out of that
position and move forward.

The following aim so to do that by using an error type all the way from
state and catching that error code in the charm deployer.

Interestingly the state tests don't change as we're backwards compatible
with the old error message (nice!) and the charm one isn't exactly easy
to trigger.