Improve ec2 retry message when waiting for root disk to become available
This commit fixes LP1915057 by rephrasing the message emitted by juju
when the instance is not yet available (EC2 resources are eventually
consistent) and we need to retry. In addition, that message is now logged
at DEBUG level and any other type of error (e.g. expired credentials
etc.) gets immediately surfaced rather than waiting for the retry
attempts to be exhausted.