Ensure we use the same user-agent version
The user-agent we send from Juju differs from client to client. The
following starts a trend to change that, so that we're consistent with
how we're querying the APIs.

This becomes apparent when the charmstore shim API has a different
User-Agent version from the same Juju version. This is a bit weird and
inconsistent.

Fixes: https://bugs.launchpad.net/juju/+bug/1913754