Improve remove-cloud output when the cloud is not known to controller
Prior to this change, doing a `juju remove-cloud bla` for a cloud known
only to the client would show a success message when removing from the
client and an ERROR message when attempting to remove from the
controller. The command itself would then exit with status 1 even though
no real error actually occurred (the controller returned a NotFound
error for the cloud).