Add list charm resources
The following adds the ability to list a charms resources from a
charmhub perspective. The facade implementation is rather easy as we can
reuse a lot of existing functionality.

I did spot an issue with the implementation, the core/charm repository
interface was importing params, which is a massive no-no. I fixed that
to have a much cleaner implementation and removed a ton of code as well.

With that fixed the list charm resources works as designed.