Re-write multi-model test in bash
The multi-model test in python would fail because it would loose track
of which model to query. Considering we're attempting to remove thses
python tests in favour of the bash ones, I took the time to do that.

The test is actually really easy and just creates a series of models and
ensures you can switch and talk to each one independantly.