Revise juju help text
This commit revises the juju help text to make it more concise and to
ensure that it prominently displays k8s-related information.

The starter command list has been re-ordered to closely match the
commands that new juju users will be working with and the REPL help text
has been shorted (the prompt text suffices for letting users know how to
work with the shell) and embedded into the help text.