Find with options
The options with find allow us to express more filter options for
querying the find api in a typed way.

We could resort to using reflect to walk the options struct, but it
makes for a more complicated code base which is harder to test and use.
Instead for now, we can use duplication and just test accordingly.