Find filter cmd
Update the find command to take into consideration the filter logic.
Unfortunately there are some filter logic that are present in the api
doc that aren't currently supported, although they will be added and
some are only available on staging api.

The following will be added, but are still being implemented

 - platforms
 - relation-requires
 - relation-provides

The following is in staging

 - publisher

I'm going to leave the ones to be implemented on the server, but hide
them in the cmd, so we can just add them when they become available.

---

The find code now removes the bespoke find logic and so we do no longer
inherit from the charmhub command. This makes the command more bespoke,
but that's a reality of what the command actually does.