Filter non-bridge networks during subnet discovery on lxd provider
When we use lxd as a substrate (vs deploying an lxd workload on a
machine managed by another substrate), Subnets() output should only
include networks that are able to be bridged.

Given that an lxd server may be either local or remote, we cannot
guarantee that we will be able to create bridges for regular NICs (as we
do in the containerizer's bridge policy logic). Therefore, by ensuring
that all reported subnets can be bridged, we allow users to define
spaces and guarantee that the lxd provider will always be able to use
the bridge NICs as parents of devices that get injected into the
container specs constructed by StartInstance.