Prevent charm series and systems (bases)
The following brings in the charm change that prevents series and
systems being used at the same time. We should aim for conceptual purity
with in the modeling of our systems and having series and systems
prevent that.

So that `series: [focal]` could deviate from `bases: {name: ubuntu,
channel: "16.04/stable"}`. This could cause chaos in the future and we
should prevent it from ever happening.

----

Hyrum's Law states:

> With a sufficient number of users of an API,
> it does not matter what you promise in the contract:
> all observable behaviors of your system
> will be depended on by somebody.