Extract upgrade series logic
This is the first step at extracting the logic from state into the API
command pattern. It attempts to push down the business logic below the
facade level and push up the logic from state. That way it can be better
tested.

With this change we can start to implement (in following PRs) the
ability to query the charmhub API when validating the applications for
the correct series.