Verify application against API
The following introduces the validation process against the charmhub
API. We use two different validators to check if an upgrade series is
allow via an application. Previously this was all done in state and now
we use the charmhub API to verify that for us.

A model can contain two different type of charms, a state charm (local
and charmstore charm) and a charmhub charm. To ensure that we don't
create a heterogeneous deployment we use a check to ensure that we call
the right one.