Ensure we wait for ha
Although there are two loops to check for machines and then the ha
status, you have to do it this way, as you want to wait for the machines
to start and then check they're in a ha-state.

We could just wait for the ha to reach 3 no matter what, but it leads to
a bad debugging experience.