Initial commit of adding tests for install commands
As we don't have any tests for install commands inside the snap package
it would be nice to do so here.

There was a bug in confinement where it would assume to use the
application confinement for the prerequisite application confinement.
They're not the same thing at all and this changes that.