Introduction of a upgrade series FSM
The previous logic for upgrade series was really bad, essentially it
allowed you to jump up the order to reach the error state without
actually ensuring that you could access all the states.

The following code changes that and introduces a FSM. The FSM ensures
that you can go to only the states that are supplied in the child
states.

The FSM is modelled as a DAG with each state show how it can move along
the edges to each vertex. This tightens up a lot of the logic around a
upgrade series and ensures that you have to move to each state
correctly.

Although this might seem overkill it should show that it's vitially
important for an upgrade series to walk over the correct states
correctly.