Introduction of validation step into FSM
The following introduces the validation step into the upgrade series
FSM. It handles the idea that an upgrade series prepare started
shouldn't happen unless a validation has happened with the lock in
place.

This basically introduces a weak transaction and rollback semantics for
upgrade series. The client can still validate out of band, but will be
removed in future commits. This will ensure that someone attempting an
upgrade series will flag all units of a machine with a upgrade series
lock and truly make Juju, multple user aware.

We're still only half way into this and there is probably more of a
rabbit hole to dig, but for now, we have a good set of semantics and
abstractions for this to correctly work.

---

Note: that in order to support backwards compatible client/controller
symantics we allow the following:

  - Not started -> prepare started
  - Not started -> validation
  - Validation -> prepare started

This should indeed make validation truly optional for the uniter when
doing upgrade series. We can probably enforce the validation step in 3.0
though and worth a thought.