Use the status API to get the units affected
The previous way to get the units was to call validate and validate
would return a list of units. The problem with this is that validate
was doing too much. It should validate and not return the units, it
cross cutting concerns. Instead another call or renaming the call to
something else would have been better.

The code here is simple, walk over the status result, returning the
units for a given machine.