Initial work for update machine series
During a finish upgrade series we ensure again that the applications are
valid, but this is superfluous. The machine lock was created and
validated upfront and we know for certain that the application supports
the given series. We therfore don't need to perform the request again.

The following code starts to extract the mess and performing a
FinishUpgradeSeries will just use the machine units, which include
subordinates before setting the series.

Interestingly performing a complete on an upgrade series doesn't
actually require you to do an upgrade, instead it will just finish with
the hostSeries, essentially providing a MASSIVE no-op.