Validate state and charmhub charms
The following validates state and charmhub charms based on the local and
remote states. Charmstore and local charms are validated against state
and charmhub charms are validated against the remote state.

The code is similar but not identical to the machine manager facade
implementation, yet is different enough to warrant these changes.