Force "ubuntu" release when building juju version in tests
All tests that upload fake tools assume "ubuntu" as the release. Prior
to this change testing.CurrentVersion auto-detected the host that the
tests were running on and used that as the Version.Release field. As a
result, all these tests would break when running on non-ubuntu systems
(e.g. centos).