Make tests that depend a particular host OS to be OS-independent
This set of changes provides overrides for code that invokes the
HostOSTypeName helper from the os package and updates the relevant tests
accordingly so that their behavior does not depend on the OS where the
tests are actually running on. This change is required to allow the
tests to pass on non-ubuntu hosts (e.g. centos).