Use charm url name when migrating charms.
During a migration it is expected that a charm name in the metadata is
the same one in the URL. This isn't always the case. In-fact it might
never be the case for CharmHub charms, as the charmhub store does a lot
of charm url re-writing.

To this end, we allow the passing in of the charm name during a
migration and fallback to the one in metadata if it's empty.

---

The code is TERRIBLE in this file. There is no logical separation, just
massive freaking function of code. It's impossible to test at a unit
level and as kat-co pointed out years ago this handler setup should be
broken up.

I've added my 2-cents to the "fixme" at the top of the file and we could
use something like a Middleware pattern for this, so we can test the
various pathways to have a better and more secure API server.

---

Fixes: LP#1921722