Calculate network CIDR from observed addrs when populating subnet details
When using netlink as our address source, it reports interface addresses
as IPNet values which represent the address in CIDR notation, e.g.
10.1.2.99/24.

Prior to this change, on manual machines, the networkingcommon package
used the address in CIDR notation value as the subnet's CIDR. API calls to
SetMachineAddresses (machiner) first convert the input arguments into a
SpaceAddresses list. However, that particular code path eventually calls
core/network/SubnetInfos.GetByUnderlayCIDR with the incorrect subnet
CIDR which fails validation thus causing an error to be returned back to
the machiner worker which in turn causes it to restart.