Upgrade charm-url to include bases
The following adds an upgrade step to ensure that previous charm urls
are re-written from series to bases. This probably breaks a lot of
existing things, as deployment can no longer happen, as it's expected
that series is a thing.

The code is quite simple, it just walks over the charms and application
collections.

FYI: this works as well for all 2.9 RCs, as a build tag is stripped and
matched with the major,  minor and patch. So 2.9rc9 is matched to 2.9.