Use Base in Charmhub package and its response structures.
Make the least intrusive changes from platform to base while
allowing for base to be used when communicating with the Charmhub
API. This is a first step to replacement within the entire juju
code base.

Add RefreshBase to correspond with RefreshPlatform.  All translation
from RefreshPlatform to RefreshBase happens within the charmhub package.

Translate from a Base in the Charmhub API response structs inside of the
charms and charmhub facades.