Distinguish between missing and empty link-local attr in netplan configs
According to the netplan reference https://netplan.io/reference/, for
the link-local field:

"If this field is not defined, the default is to enable only
IPv6 link-local addresses. If the field is defined but configured as an
empty set, IPv6 link-local addresses are disabled as well as IPv4 link-
local addresses."

This commit changes the type of the LinkLocal field in netplan.Interface
to a *[]string so we can distinguish between a missing link-local
attribute and a present link-local attribute with an empty list as its
value.