Tail the current log
One issue we have currently is when developing locally it's hard to know
what's being run and where the error arises. The following changes that
by launching a background process that tails the current test log and
just renders that to stdout/stderr if you have -v enabled.

The good thing with this, is that we get to keep backwards compatibility
by ensuring that we keep the test lof file around for archiving.