Help when no controller is selected
One thing that annoys me is that we have all the information at hand to
help with the UX for simple error messages and a path forward to users.

A case in point is that when you end up having removed or destroyed a
controller and the controllers.yaml isn't correctly updated, then you
get a brain dead error message.

The code fixes this and gives a pointer to the user about what they
could select, without having to go to another command `juju controllers`
to get that information.