Allow base channel to also contain a risk
The risk is there to be used sometime in the future, we don't know how
to handle it atm, but we shouldn't error out when we see it. The
following allows the parsing of that.

I expect the future implementation to just drop the risk on the floor,
but we shouldn't prevent charmhub charms from dying because of it.