Ensure we push bases all the way through
Although the API still accepts platform for refresh API, it's actually
not the right way to do it. Instead we should push bases all the way
through the stack. That way we know that this is actually future proofed
against any API changes.

The annihilation of platfrom from charmhub client.