Ensure we validate the model target
Instead of using the charm URL, which will be wrong for bases, instead
use only the metadata. The metadata will always be right!

So the following code changes what the purpose of this validate code,
from catching a non container on a container work load. We drop the
series check because we validate that check in other places.