k8s charm url might not have a series
The k8s charm url might not have the k8s url in it during a set charm,
if that's the case we need to check the meta data to ensure that it
actually is.

I've remove the if from the if/else as I think we should always check
it, but I might be wrong there?