CAAS bundle deployments
As we don't have the kubernetes in the charm URL we have to then
fallback to the charm metadata. In this instances we can do that and we
can inspect it.

The great thing about this change is that is suppliments the existing
workflow of checking the computed series, so this should work with old
style charms. The worst case is that you might end up with checking the
same series twice, but that's always been the case.