V2 Format charm
The following is a v2 format charm that isn't k8s specific. That way we
can ensure that you can deploy a v2 version of the charm that follows
the manifest requirement.

It is expected that you have charmcraft installed via

    $ snap install charmcraft --channel=edge

Then cd in to testcharms/charm-hub/charms/juju-qa-test-v2 and run

    $ charmcraft build
    $ juju deploy juju-qa-test-v2.charm