Series-less charms
The following program attempts to locate series-less charms on
charmhub. These charms will not have a series or a map of containers.

This script is useful to identify problem charms that made the charmhub
cut. These charms are either old or should be updated to have a series
with in the metadata.

The code is nothing special, just brute force the API.