Improve error message for invalid API result
It is expected that we can always get valid API results, except when we
don't we should return an error message that is more informative. In
this instance we end up returning an error message about missing bases.

The following code just improves that by giving a bit more context. I
can't change the error type as we use that to diagnose other issues, but
I do wish that `not found` wouldn't be appended to the end of the error
message as it's rather restricting on the message returned.