Handle internal server errors from charmhub
Server errors from charmhub can return back plain/text mime types, even
though we tell the server we only accept json. To better improve the
error message we check the status code and return back if it's a 5xx
error.

The user never has to see the dump from the server and we can get the
dump from the response in the server logs.

We also ensure that we close the body and allow the response to be
reused after draining.

----

As a driveby we introduce a MIME type alias because it's painful to see
"application/json" everywhere.