Handle k8s v1 charm inside a bundle
Deploying a v1 charm metadata inside a k8s bundle looses the ability to
know that the computed series for a charm is actually a kubernetes
charm.

The code just checks to ensure that the computed series also contains
kubernetes so that we can deploy correctly. This is a workaround for v1
charm metadata and if we ever remove support for it, we can safely drop
kubernetes series.