Introduce better testing for the deployer
The deployer code isn't very well tested, this starts to change this. By
adding an indirection to the ReadCharm function call, we can now add
smaller unit tests to expose the error we're after.

Implementing a temporary charm reader will still exercise the more
integration side of the testing, so we're still covering the bases
there.