Prevent starvation when asking for random values
The previous requesting of a 100M file would fail because the head
couldn't get enough characters to fullfill the request and die. Weirdly
you can get it pass sometimes and not others. The fix for this is to
request a large enough size file without the usage of head. Using dd
works nicely here and we just push it through base64 so we get a clean
output in juju status.