Select default charm URL schema depending on controller version
Prior to this change, 2.9 clients always assumed charmhub (ch) to be the
default charm URL schema for charms URLs without one. This breaks
backward compatibility with older controllers when a 2.9 client attempts
to deploy a charm (`juju deploy wordpress`) without an explicit schema
prefix to a 2.8 controller.

In this case, the client assumes that we try to deploy a charmhub charm to the
controller but since the controller lacks the appropriate APIs, the
charm API client returns back a NotSupported error which the charm
adapter spits out to the user with a totally incomprehensible error:
"ERROR resolve charms not supported".

To address this problem, the client will first check the supported charm
API version and use this information to select the default schema for
schema-less charm URLs. This is accomplished via the updated
charm/v8.EnsureSchema signature which requires the default schema to be
specified.

The default schema is threaded through the various deploy-related
structs (charms/bundles etc.) and is set to charm.CharmHub when dealing
with refresh requests as these are only valid against 2.9 controllers.

Finally, if the user tries to explicitly deploy a charmhub charm against
a pre 2.9 controller, they will get a straightforward error with a
recommendation to update to 2.9+ if they want to use charmhub.