Relax origin validation ruls when resolving charms due to a charm switch
The ResolveCharm API call payload has been augmented with an optional
"SwitchCharm" field which the client should set when resolving a charm
prior to a 'juju refresh --switch' invocation. In this case, the
controller should not return an error in case of an (original) origin /
charm schema mismatch as switching should allow the user to change from
a charmstore charm into the charmhub equivalent.