CLI: Improve error message for existing relations
If the removal of relation fails and the user of juju isn't using
`juju status --relations` then they wouldn't notice that the relation
was in error and when they tried to re-add the relation you'd get a
cryptic error that it already exists.

The solution is to just work out if the endpoint is error using the full
status and improve the error message there.