Fix the feature test to regcognise we error out
The feature test was using the fact that we never error'd out for a
add-relation if it already existed. That was disingenuous to the user as
the transaction never ran.

The work around for people using a CLI to drive things from a bash
script or similar is to `|| true` or inspect the output of stderr.