Clean up stale single-peer relations when switching charms
Juju allows switching to a charm that would otherwise break an existing
peer relation as long as there is only one unit present. However, doing
so leaves stale documents in the relations collection.

As a result, if deployed a single-unit application, then switched to a
charm without peer relations and then tried to switch back to a charm
with peer relations we would get in a state where the SetCharm call
would attempt to insert a new document into the relations collection but
its DocMissing assertion would fail since the doc was still there. As a
result the txn runner would eventually give up and return a
"contention error".

To address this, this commit first modifies the
check-if-relations-will-be-broken logic so that it allows us to change
the charm if the application has zero OR one units. Prior to the change
we only allowed the change to proceed if there was only one unit
present. However, you can still have applications with no units and that
shouldn't block the switch.

Secondly, an extra check has been added to the changeCharmOps method
which performs a diff between the existing peer relations and the peer
relations that the new charm defines and removes any stale relations
that were previously left behind after the switch.