Locate container and machine units regression
The following was a regression when attempting to locate the right units
for a given container. Only machines where checked which caused issues
when attempting to upgrade a container based instance.

The fix is really easy, as it just looks through the machine containers
as well.

It is expected that performing an upgrade series, the model and in turn
the status has quiesced before proceeding.