Facades: Update charm facade to consume all bases
The following changes allows us to collate all bases when no base is
passed. This is possible because of a change to charmhub API; passing
multiple bases back within the error.

Previously the code would select the first base within the error and use
that for the resolution of the charm, dropping any others on the floor.
It wasn't required to do anything more with them at the time.
The change here is that we collate them, so once we've resolved the name
against the API, we can extract the series from them and return them
back. Allowing the client side logic to see if the model-default value
is set on it, if it is, we can use that for the actual charm download
selection.

The code changes now expect the result from the selection from the error
to return many and not 1. Additional changes to simplify the setting of
the original input charm origin is also used, which makes reading the
code much easier.

Additionally the way this is implemented, this will work in a backwards
compatible way, as the underlying API hasn't changed the type, just how
much data is returned.