Charmhub: Improve download logging on failure
The following changes aims to improve debugging when attempting to
download a charm. Unfortunately we can end up with server errors that
means that a download of an archive will fail. In order to not leak the
client implementation out to the user, we hide the error status code,
but with this logging, a user can use `juju debug-log` to at least
inspect it if it's required/persisent.