Facades: Improve retry readability
The following attempts to improve the retry readability of the resolving
of a charm against the charmhub API.

The code expects that in most cases we have to retry with different
synmantics and doing so should be done carefully. We don't want to
create a generic retry around this, as that will confuse people about
what is actually happening. Instead we unroll the loop and and correctly
retry with different arguments.