GCE: Allow inject of global http client
The following code changes starts to lift the http client out from the
edge and into the constructor of the environ. Future commits should
allow us to pass in a http client on the construction of the environ
from the dependency engine. This is just the first step of many to do
that.