Simplestreams: Allow test clients
The following removes the testing hack to prevent out going requests and
simply uses a different client for tests. This means injecting a
different data source factory, which creates a different client based on
tests.

The code is very simple and it means we start to understand our
dependencies inside of simplestreams a bit better.

---

In reality the real problem here is that we should be creating a
simplestreams type that can share that knowledge in a much simpler
fashion.