Integration tests: Server IP address
The server IP address for the metadata service has to be routable, but
not the loopback one to work in every scenario. In this case we just
make this more robust by iterating over the `hostname -I` command and
finding the first one that works correctly.

This should be enough for juju to correctly bootstrap from a local file
system.