Openstack: Use juju/http client
The following updates the openstack client factory to use the juju/http
client instead of the one offered by the library. We can reduce a lot of
the client boiler plate by just surfacing the only client.

More work is required to lift it out of the provider into a manifold,
but that work can be done at a later PR.