EC2: Provide comments with context.
The comment is useful just as a commit message:

This isn't in it's final form. It is expected that eventually the ec2
client will be passed in via the constructor of the environ. That can
then be passed in via the environProvider. Unfortunately the provider
(factory) is registered in an init function and makes it VERY hard to
override. The solution to all of this is to remove the global registry
and construct that within the main (or provide sane defaults). The
provider/all package can then be removed and the black magic for
provider registration can then vanish and plain old dependency
management can then be used.