LeadershiptContext: Embed a leadership context
The following is a very simple mechanical patch with the aim of updating
the facade context in the future. The facade context should be a gateway
to types/objects.

The context becomes:

type Context interface {
    Leadership() Leadership
}

type Leadership interface {

    ...
}

Although it's close to what we currently have, it becomes more of a
single responsibility proposal.