Charmhub: Better request logging
The following ensures that we have single responsibility for logging
when performing API requests.

The default HTTP transport will also give you a logging request
recorder, this should aid with always having a request recorder for
making charmhub requests.

As all the above is handled only when trace is enabled shouldn't impact
performance.