Metrics: Wireup total requests api metrics
The following wires up the outbound total requests API metrics. It uses
the juju/http request recorder.

We have to be extra careful about what we want to use as labels, we
don't want to log the path or the complete URL for a couple of reasons:

 1. Security around what can be in a URL; auth tokens et al
 2. We don't want to balloon the cardinality of the database.

With this in mind, we want to just request the host names for now, which
keeps the cardinality low.