Facades: Improve dependency management for newRoot
The newRoot constructor took a various amount of different root types.
Instead of this, we just make apiHandler conform to an interface and
then we can just select the types off of that.

The test version was also easy to construct, just create an empty type
and pass back nil in most cases or a dummy resource container.

This should make the readability of a newRoot type easier to understand.