Support ssh/scp --proxy for containers with FAN addresses
The juju ssh --proxy command uses the controller as a jumpbox and
invokes 'nc' to forward ssh traffic to the target machine. While this
approach works fine for regular machines, it fails (for substrates as
ec2) for container machines (lxd/kvm) when they only have a FAN address.

This is generally acceptable as juju only requires for machines to be
able to connect to the controller and not vice-versa. In particular, ec2
does not allow FAN connectivity from the controller model to any other
Juju model (and vice versa).

To workaround this limitation, we now first check whether the machine we
are trying to connect to is a container. If that's the case, we identify
the address of the machine that hosts the container and modify the ssh
proxy command to use a 2-level proxying scheme: we ssh to the
controller, then ssh to the machine hosting the container and then run
'nc' to route the ssh traffic to the intended machine.