Add -c <cloud-name> to test options.
Allow users to test against clouds with different names from the
provider or cloud type, e.g. vsphere, openstack.  Introduces a
BOOTSTRAP_CLOUD, which is different from the BOOSTRAP_PROVIDER.

The "microk8s" provider is now the "k8s" provider.  "microk8s" would be
a cloud.  This allows other k8s clouds to be tested against.

vSphere and OpenStack are now supported as providers.  Use with the -c
option to provide their cloud name local to the environment.

vSphere requires use the juju boston vsphere for now.  Embedded in the
bootstrap is setup of vm templates for reuse, rather than creating new
ones for up and download, making the tests run faster.

OpenStack requires images to be done automagically, rather than
specifying metadata on the bootstrap cli.

Do not try to introspect a k8s controller, the commands do not work.