Action: Task exit code
The following propagates the task result exit code up to the CLI so that
a user calling `juju exec` will know that an error occurred in one or
more tasks. Prior to this commit, users would have to inspect every
result which is a bit disappointing for bash style scripting against the
CLI. With this change it should be possible to do the following:

   $ juju exec --all 'exit 1'
   $ echo $? # should see 1

This makes juju a lot more powerful to sys or ops users.

The code is simple, inspect the result for an exit code. If it exists
propagate it to the CLI and return an error stating the fact that a task
failed.