Cache: Cache coherence
The cache would fail to notify if you changed from a value and back to
the original value. This is because the underlying hash value was never
written. This has two consequences:

 1. The cache could never emit a change of the original value
 2. Because we never cached the new value, if the new value was the same
 as the old one, we would still emit the event.

The fix is easy, just save the hash after the notify and we can have
better cache coherency.