Improve transactions in action removeAndLogBuildTxn.
Related to LP:1931567.  1) When enqueuing an operation, include the expected
task count to start with.  The operations task count may not be the
expected count and can change while an operation is running.  2) Add an
assertion to check the current count is 1 less than expected count
before marking an operation complete.