Deploy: Deloyment of private charms
When resolving a private charm we need to correctly discharge a
macaroon, so that we can get a token. When we moved the resolving of
charms to the server side, this was taken into account. In the end we
believed that it would be better to fallback to the previous method of
charm resolution. This is until server side deployments will be a thing.

The code path for this in the client was wrong. The strategy for charm
resolution states; if there is no error, then just return. Else if it's
a charmhub charm and it's not supported show a nice error, otherwise
just fail. If it's any other charm, fallback to the prior workflow. This
should cover all the scenarios.