State logs: index label fields
When using debug-logs it helps if we index the log fields correctly. The
following change ensures that we index the labels field (in this
instance it's the "c" field). As the field is optional, we now need to
create a sparse index for mongo.

See: https://docs.mongodb.com/manual/core/index-sparse/