Deploy: Allow v1 kubernetes deployments
The following ensures that we can deploy v1 charms that resolve with
focal API resolution. This broke when cleaning up how we handle series
(for good), but we accidentally broke v1 charms for kubernetes.

The solution is check for kubernetes charms and v1 format, force the
charm series to kuberenetes.

Replication is easy:

    cat <<EOF >bundle.yaml
    bundle: kubernetes
    applications:
      prometheus:
        charm: prometheus-k8s
        scale: 2
      grafana:
        charm: grafana-k8s
        scale: 1
    relations:
      - - prometheus
        - grafana
    EOF

    juju bootstrap microk8s
    juju deploy bundle.yaml

Signed-off-by: Simon Richardson <simon.richardson@canonical.com>