Scripts: Juju inspect
Juju inspect attempts to improve the live debugging of a live juju
controller (HA). This is just in it's infancy and is missing a lot of
knowledge. The was originally formulated whilst trawling through 100s of
engine reports and loosing track of what was what.

This was originally trying to workout what was the mongo primary, but
found uses in lots of other ways. Restart counts, raft issues et al.