Inspect: Generic rules
The rules are now more generic, ideally we could use something like
json.RawMessage, but there isn't a yaml.RawMessage. The work around is
to just unmarshall to an interface and then re-marshal back into
something more specific when required. As we're not looking for pure
speed here, we can safely do this. Hopefully we can then make more rules
without having to generalise the worker type.

Additionally cleaned up the logging so the rules can write how ever they
like to the writer.